mailrun now monitors offlineimap differently
Before this patch we would check to see if offlineimap was running and
kill it if so, assuming it's a hung process. This led to a few occasions
where mailrun killed a manual sync I was doing from mutt.

This patch changes mailrun in stead babysit offlineimap for a minute,
killing it if it hasn't yet completed.

Now that we know an existing process can only come from a true sync, we
can exit ourselves when that's encountered and not kill things we
shouldn't.